feat: rename @dfinity/icp to @icp-sdk/core
