+++ e2e/node/basic/trap.test.ts
@@ -5,10 +5,10 @@ import {
   HttpAgent,
   AgentError,
   CertifiedRejectErrorCode,
-} from '@dfinity/icp/agent';
+} from '@icp-sdk/core/agent';
 import util from 'util';
 import exec from 'child_process';
-import { IDL } from '@dfinity/icp/candid';
+import { IDL } from '@icp-sdk/core/candid';
 const execAsync = util.promisify(exec.exec);
 
 let stdout;
