+++ e2e/node/canisters/declarations/counter/index.d.ts
@@ -1,6 +1,6 @@
 import type { ActorSubclass, HttpAgentOptions, ActorConfig, Agent, Actor } from '@dfinity/agent';
 import type { Principal } from '@dfinity/principal';
-import type { IDL } from '@dfinity/icp/candid';
+import type { IDL } from '@icp-sdk/core/candid';
 
 import { type _SERVICE } from './counter.did';
