+++ e2e/node/tsconfig.json
@@ -11,7 +11,7 @@
   "include": ["utils/**/*.ts", "canisters/**/*.ts", "basic/**/*.ts"],
   "exclude": [".dfx", "lib", "node_modules"],
   "references": [
-    { "path": "../../packages/icp" },
+    { "path": "../../packages/core" },
     { "path": "../../packages/agent" }
   ]
 }
