+++ e2e/node/utils/canisterid.ts
@@ -1,4 +1,4 @@
-import { Principal } from '@dfinity/icp/principal';
+import { Principal } from '@icp-sdk/core/principal';
 import { execSync } from 'child_process';
 
 export const getCanisterId = (canisterName: string): Principal => {
