+++ packages/core/.npmignore
