+++ packages/core/README.md
@@ -1,31 +1,31 @@
-# @dfinity/icp
+# @icp-sdk/core
 
-The `@dfinity/icp` package is a JavaScript and TypeScript library designed to interact with the Internet Computer. It acts as a meta-package that re-exports functionality from several other `@dfinity` packages, providing a convenient way to access multiple tools and utilities in one place.
+The `@icp-sdk/core` package is a JavaScript and TypeScript library designed to interact with the Internet Computer. It acts as a meta-package that re-exports functionality from several other `@dfinity` packages, providing a convenient way to access multiple tools and utilities in one place.
 
 **Note:** This package requires the `node16` (or later) module resolution strategy to function correctly.
 
 ## Installation
 
-You can install the `@dfinity/icp` package using your preferred package manager:
+You can install the `@icp-sdk/core` package using your preferred package manager:
 
 ### npm
 ```bash
-npm install @dfinity/icp
+npm install @icp-sdk/core
 ```
 
 ### yarn
 ```bash
-yarn add @dfinity/icp
+yarn add @icp-sdk/core
 ```
 
 ### pnpm
 ```bash
-pnpm add @dfinity/icp
+pnpm add @icp-sdk/core
 ```
 
 ## Package Structure
 
-The `@dfinity/icp` package re-exports modules from the following `@dfinity` packages:
+The `@icp-sdk/core` package re-exports modules from the following `@dfinity` packages:
 
 - [`@dfinity/agent`](../agent/README.md): Provides tools to interact with the Internet Computer, including creating actors and making calls.
 - [`@dfinity/assets`](../assets/README.md): Utilities for managing assets on the Internet Computer.
@@ -39,24 +39,24 @@ The `@dfinity/icp` package re-exports modules from the following `@dfinity` pack
 
 The package provides the following import paths for accessing the re-exported modules:
 
-- `@dfinity/icp/agent`
-- `@dfinity/icp/assets`
-- `@dfinity/icp/auth-client`
-- `@dfinity/icp/candid`
-- `@dfinity/icp/identity`
-- `@dfinity/icp/identity-secp256k1`
-- `@dfinity/icp/principal`
+- `@icp-sdk/core/agent`
+- `@icp-sdk/core/assets`
+- `@icp-sdk/core/auth-client`
+- `@icp-sdk/core/candid`
+- `@icp-sdk/core/identity`
+- `@icp-sdk/core/identity-secp256k1`
+- `@icp-sdk/core/principal`
 
 ## Usage Example
 
-Here is an example of how to use the `@dfinity/icp` package:
+Here is an example of how to use the `@icp-sdk/core` package:
 
 ```typescript
 // Importing the HttpAgent from the agent module
-import { HttpAgent } from '@dfinity/icp/agent';
+import { HttpAgent } from '@icp-sdk/core/agent';
 
 // Importing the Principal utility
-import { Principal } from '@dfinity/icp/principal';
+import { Principal } from '@icp-sdk/core/principal';
 
 // Create an HttpAgent instance
 const agent = new HttpAgent({ host: 'https://ic0.app' });
