+++ packages/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dfinity/icp",
+  "name": "@icp-sdk/core",
   "version": "1.0.0-beta.0",
   "author": "DFINITY Stiftung <sdk@dfinity.org>",
   "license": "Apache-2.0",
@@ -8,7 +8,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/dfinity/agent-js.git",
-    "directory": "packages/icp"
+    "directory": "packages/core"
   },
   "bugs": {
     "url": "https://github.com/dfinity/agent-js/issues"
