+++ packages/core/src/assets/index.ts
