+++ packages/core/src/identity-secp256k1/index.ts
