+++ packages/core/tsconfig.cjs.json
